+++ src/main/java/com/magic/project/services/implementation/PatientServiceImplementation.java
@@ -7,6 +7,7 @@ import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.magic.project.handler.PatientNotFoundException;
 import com.magic.project.models.Patient;
 import com.magic.project.repositories.PatientRepository;
 import com.magic.project.services.PatientService;
@@ -25,15 +26,19 @@ public class PatientServiceImplementation implements PatientService {
 	@Override
 	public Patient deletePatient(@Valid String patId) {
 		Patient Patient = patRepo.findById(patId).orElse(null);
-//		if (Patient==null){
-//			throw  new Exception("No Patient with ID "+patId);
-//		}
+		if (Patient == null) {
+			throw new PatientNotFoundException("No Patient with ID " + patId);
+		}
 		patRepo.deleteById(patId);
 		return Patient;
 	}
 
 	@Override
 	public Patient updatePatient(Patient updatedPatient, @Valid String patId) {
+		Patient Patient = patRepo.findById(patId).orElse(null);
+		if (Patient == null) {
+			throw new PatientNotFoundException("No Patient with ID " + patId);
+		}
 		updatedPatient.setPatId(patId);
 		patRepo.save(updatedPatient);
 		return updatedPatient;
@@ -41,7 +46,10 @@ public class PatientServiceImplementation implements PatientService {
 
 	@Override
 	public List<Patient> getPatientList() {
-		List<Patient> Patients = patRepo.findAll();
-		return Patients;
+		List<Patient> patients = patRepo.findAll();
+		if (patients.isEmpty()) {
+			throw new PatientNotFoundException("No Patient Found.");
+		}
+		return patients;
 	}
 }
