+++ src/main/java/com/magic/project/services/implementation/ReceptionistServiceImplementation.java
@@ -7,6 +7,7 @@ import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.magic.project.handler.ReceptionistNotFoundException;
 import com.magic.project.models.Receptionist;
 import com.magic.project.repositories.ReceptionistRepository;
 import com.magic.project.services.ReceptionistService;
@@ -24,16 +25,20 @@ public class ReceptionistServiceImplementation implements ReceptionistService {
 
 	@Override
 	public Receptionist deleteReceptionist(@Valid String email) {
-		Receptionist Receptionist = recepRepo.findById(email).orElse(null);
-//		if (Receptionist==null){
-//			throw  new Exception("No Receptionist with ID "+email);
-//		}
+		Receptionist receptionist = recepRepo.findById(email).orElse(null);
+		if (receptionist == null) {
+			throw new ReceptionistNotFoundException("No Receptionist with ID " + email);
+		}
 		recepRepo.deleteById(email);
-		return Receptionist;
+		return receptionist;
 	}
 
 	@Override
 	public Receptionist updateReceptionist(Receptionist updatedReceptionist, @Valid String email) {
+		Receptionist receptionist = recepRepo.findById(email).orElse(null);
+		if (receptionist == null) {
+			throw new ReceptionistNotFoundException("No Receptionist with ID " + email);
+		}
 		updatedReceptionist.setEmail(email);
 		recepRepo.save(updatedReceptionist);
 		return updatedReceptionist;
@@ -42,6 +47,9 @@ public class ReceptionistServiceImplementation implements ReceptionistService {
 	@Override
 	public List<Receptionist> getReceptionistList() {
 		List<Receptionist> receptionists = recepRepo.findAll();
+		if (receptionists == null) {
+			throw new ReceptionistNotFoundException("No Receptionist Found.");
+		}
 		return receptionists;
 	}
